misc refactors
- move client factory to the signing clients folder
- use `debug` package instead of custom debug logic to prevent logs in tests
