+++ packages/garasign-utils/package.json
@@ -66,7 +66,8 @@
     "typescript": "^5.0.4"
   },
   "dependencies": {
-    "commander": "^11.1.0",
+    "commander": "^10.0.1",
+    "debug": "^4.3.4",
     "ssh2": "^1.15.0"
   }
 }
