+++ packages/garasign-utils/src/utils.ts
@@ -1,15 +1,6 @@
-import path from 'path';
-import type { SSHClientOptions } from './ssh-client';
-import { SSHClient } from './ssh-client';
-import { LocalSigningClient, RemoteSigningClient } from './signing-clients';
-import type {
-  SigningClient,
-  ClientType,
-  ClientOptions,
-} from './signing-clients';
+import { debug as debugFn } from 'debug';
 
-// eslint-disable-next-line no-console
-export const debug = console.log;
+export const debug = debugFn('garasign-utils');
 
 export function assertRequiredVars() {
   [
@@ -22,37 +13,4 @@ export function assertRequiredVars() {
       throw new Error(`${envVar} is required`);
     }
   });
-}
-
-function getSigningScript() {
-  return path.join(__dirname, '..', 'src', './garasign.sh');
-}
-
-async function getSshClient(sshOptions: SSHClientOptions) {
-  const sshClient = new SSHClient(sshOptions);
-  await sshClient.connect();
-  return sshClient;
-}
-
-export async function getSigningClient<T extends ClientType>(
-  client: T,
-  options: ClientOptions<T>
-): Promise<SigningClient> {
-  if (client === 'remote') {
-    const sshClient = await getSshClient(options as SSHClientOptions);
-    // Currently only linux remote is supported to sign the artifacts
-    return new RemoteSigningClient(sshClient, {
-      rootDir: '~/garasign',
-      signingScript: getSigningScript(),
-    });
-  }
-  if (client === 'local') {
-    // For local client, we put everything in a tmp directory to avoid
-    // polluting the user's working directory.
-    return new LocalSigningClient({
-      rootDir: path.resolve(__dirname, '..', 'tmp'),
-      signingScript: getSigningScript(),
-    });
-  }
-  throw new Error(`Unknown client type: ${client}`);
 }
